+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 ===
 ### 0.3.3 (2017-08-26)
 * Added `Transfer-Encoding: chunked` support for HTTP requests
+* Added `--remote-trigger` option to defer printing until JavaScript function is called
 * Added HTTP headers for pdf printing options
 * Added `/list-sizes` enpoint to list available page sizes
 * Fixed bug related to `Content-Location` handling in HTTP request